refactoring/#17: categorytab 뷰로직 비지니스로직 분리
